Move TerminatorKind::Drop trans to mir::statement.
